@@ -1,4 +1,8 @@
+import os
 from setuptools import setup
+from pathlib import Path
+
+root = Path(__file__).parent
 
 d = {}
 exec(open("nwbwidgets/version.py").read(), None, d)
@@ -6,6 +10,9 @@ version = d["version"]
 
 with open("README.md") as f:
     long_description = f.read()
+
+with open(os.path.join(root, "requirements.txt")) as f:
+    requirements = f.readlines()
 
 setup(
     author="Ben Dichter",
@@ -16,31 +23,14 @@ setup(
         "Development Status :: 3 - Alpha",
         "Framework :: Jupyter",
         "License :: OSI Approved :: MIT License",
-        "Programming Language :: Python :: 3.5",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
     ],
     description="This is nwbwidgets, widgets for viewing the contents of a "
     "NWB-file in Jupyter Notebooks using ipywidgets.",
-    install_requires=[
-        "pynwb",
-        "ipympl",
-        "matplotlib",
-        "numpy",
-        "ipyvolume",
-        "ndx_grayscalevolume",
-        "plotly",
-        "scikit-image",
-        "tqdm>=4.36.0",
-        "ndx-icephys-meta",
-        "ipysheet",
-        "zarr",
-        "ccfwidget",
-        "tifffile",
-        "ndx-spectrum",
-        "trimesh",
-    ],
-    license="MIT",
+    install_requires=requirements,
+    license="BSD",
     keywords=["jupyter", "hdf5", "notebook", "nwb"],
     long_description=long_description,
     long_description_content_type="text/markdown",
@@ -51,7 +41,7 @@ setup(
         "nwbwidgets/analysis",
         "nwbwidgets/controllers",
     ],
-    python_requires=">=2.7",
+    python_requires=">=3.7",
     setup_requires=["setuptools>=38.6.0", "setuptools_scm"],
     url="https://github.com/NeurodataWithoutBorders/nwb-jupyter-widgets",
 )
