@@ -1,15 +1,18 @@
-from abc import abstractmethod
 import functools
-import matplotlib.pyplot as plt
-import numpy as np
 from bisect import bisect
-import plotly.graph_objects as go
+from abc import abstractmethod
+
+import scipy
+import numpy as np
+import matplotlib.pyplot as plt
+
 from ipywidgets import widgets, fixed, Layout
+import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 from plotly.colors import DEFAULT_PLOTLY_COLORS
+
 from pynwb import TimeSeries
 from pynwb.epoch import TimeIntervals
-import scipy
 
 from .controllers import (
     StartAndDurationController,
@@ -350,12 +353,11 @@ class SingleTracePlotlyWidget(AbstractTraceWidget):
             self.out_fig.data[0].x = get_timeseries_tt(timeseries, istart, istop)
             self.out_fig.data[0].y = list(yy)
 
+            # Get data y-range, catching case with no data in current range (if so - no update)
+            y_range = [min(yy), max(yy)] if yy.size != 0 else [None, None]
             self.out_fig.update_layout(
-                yaxis={"range": [min(yy), max(yy)], "autorange": False},
-                xaxis={
-                    "range": [min(self.out_fig.data[0].x), max(self.out_fig.data[0].x)],
-                    "autorange": False,
-                },
+                yaxis={"range": y_range, "autorange": False},
+                xaxis={"range": time_window, "autorange": False},
             )
 
         self.controls["time_window"].observe(on_change)
@@ -408,7 +410,8 @@ class SeparateTracesPlotlyWidget(AbstractTraceWidget):
                         self.out_fig.data[i].x = tt
                         self.out_fig.data[i].y = dd
                         self.out_fig.update_yaxes(
-                            range=[min(dd), max(dd)], row=i + 1, col=1
+                            range=[min(dd), max(dd)] if dd.size != 0 else [None, None],
+                            row=i + 1, col=1
                         )
                         self.out_fig.update_xaxes(
                             range=time_window, row=i + 1, col=1
@@ -690,13 +693,13 @@ class MultiTimeSeriesWidget(widgets.VBox):
             tmin=self.tmin, tmax=self.tmax
         )
 
-        widgets = [
+        ts_widgets = [
             widget_class(
                 time_series, foreign_time_window_controller=self.time_window_controller
             )
             for widget_class, time_series in zip(widget_class_list, time_series_list)
         ]
-        self.children = [self.time_window_controller] + widgets
+        self.children = [self.time_window_controller] + ts_widgets
 
 
 class AlignMultiTraceTimeSeriesByTrialsAbstract(widgets.VBox):
@@ -864,7 +867,7 @@ class AlignMultiTraceTimeSeriesByTrialsConstant(
     ):
         data, time_ts_aligned = self.align_data(start_label, before, after, index)
         if group_inds is None:
-            group_inds = np.zeros(len(self.trials), dtype=np.int)
+            group_inds = np.zeros(len(self.trials), dtype=int)
         if align_to_zero:
             for trial_no in order:
                 data_zero_id = bisect(time_ts_aligned[trial_no], 0)
@@ -955,7 +958,7 @@ class AlignMultiTraceTimeSeriesByTrialsVariable(
 
         data,time_ts_aligned = self.align_data(start_label,before,after,index)
         if group_inds is None:
-            group_inds = np.zeros(len(self.trials), dtype=np.int)
+            group_inds = np.zeros(len(self.trials), dtype=int)
         if align_to_zero:
             for trial_no in order:
                 data_zero_id = bisect(time_ts_aligned[trial_no], 0)
