@@ -1,7 +1,9 @@
 from typing import Iterable
 
-import ipywidgets as widgets
 import numpy as np
+
+import ipywidgets as widgets
+
 from hdmf.common import DynamicTable
 from pynwb.misc import Units
 
@@ -48,8 +50,7 @@ class AllenRasterGroupAndSortController(GroupAndSortController):
             x
             for x in self.electrodes.colnames
             if not (
-                isinstance(self.electrodes[x][0], Iterable)
-                or isinstance(self.electrodes[x][0], str)
+                isinstance(self.electrodes[x][0], (Iterable, str))
             )
         ]
         return units_orderable_cols + [
