@@ -1,6 +1,8 @@
-import matplotlib.pyplot as plt
 import numpy as np
+import matplotlib.pyplot as plt
+
 from ipywidgets import widgets
+
 from ndx_spectrum import Spectrum
 
 
