@@ -1,8 +1,10 @@
-from pynwb.core import DynamicTable
 import numbers
+from typing import Iterable
+
 import numpy as np
 import pandas as pd
-from typing import Iterable
+
+from pynwb.core import DynamicTable
 
 
 def extract_data_from_intervals(dynamic_table: DynamicTable):
@@ -74,22 +76,21 @@ def infer_columns_to_plot(dynamic_table: DynamicTable):
     column_names_to_plot: list
         Columns that can be plotted with the dynamic table summary
     """
-    categorical_cols = infer_categorical_columns(dynamic_table)
-    
+
     column_names_to_plot = []
-    
+
     df = dynamic_table.to_dataframe()
     categorical_columns = []
     for name in dynamic_table.colnames:
         # if name not in categorical_cols.keys():  # categorical columns can always be plotted
         value = df[name].values[0]
-        
+
         if isinstance(value, (int, float, np.integer)):
             column_names_to_plot.append(name)
         elif isinstance(value, (str, bool, bytes)):
             column_names_to_plot.append(name)
             categorical_columns.append(name)
-        
+
     return column_names_to_plot, categorical_columns
 
 
