@@ -1,5 +1,6 @@
-import matplotlib.pyplot as plt
 import numpy as np
+import matplotlib.pyplot as plt
+
 from nwbwidgets.utils.mpl import create_big_ax, grid_sps
 
 
